Add CaveEnemy and MedPack classes
Introduced CaveEnemy and MedPack classes with corresponding headers and implementations. Updated Enemy and Item base classes to support new features (float speed, item usage, destroy after use). Modified Player to support item usage and healing, and Scene to use new classes. Project files updated to include new sources and headers. Added geodude.png sprite for CaveEnemy.

+++ MetalGear/Enemy.h
@@ -32,7 +32,7 @@ public:
 	bool seePlayer(const glm::ivec2& playerPos) const;
 
 protected:
-	glm::ivec2 posEnemy;
+	glm::vec2 posEnemy;
 	glm::ivec2 tileMapDispl;
 	glm::ivec2 size;
 	vector<int> patrolSquare;
@@ -42,7 +42,7 @@ protected:
 	Player* player;
 	int health;
 	int damage;
-	int speed;
+	float speed;
 	bool playerHasBeenDetected;
 
 	Texture spritesheet;
